+++ modules/app/src/main/res/values/strings.xml
@@ -498,6 +498,11 @@
     <string name="sharedCollectionUnsubscribed">Collection successfully unsubscribed</string>
     <string name="sharedCollectionUpdated">Collection successfully updated</string>
 
+    <plurals name="sharedCollectionChangedNotificationBigMsg">
+        <item quantity="one">The collection \'%s\' has been updated with a new application. Click here to add it to your collection.</item>
+        <item quantity="other">The collection \'%s\' has been updated with new applications. Click here to add them to your collection.</item>
+    </plurals>
+
     <!-- Edit Widgets -->
     <string name="editingWidgets">Editing widgets</string>
     <string name="movingWidgets">Moving widgets</string>
